+++ src/components/ButtonRow/index.js
@@ -1,11 +1,50 @@
 import React from 'react';
+import classNames from 'classnames';
+
+const areAnyButtonsFullWidth = (buttons) => {
+  let includesFullWidth = false;
+
+  buttons.map((button) => {
+    if (button.props && button.props.width === 'full') {
+      includesFullWidth = true;
+    }
+  });
+
+  return includesFullWidth;
+};
 
 const ButtonRowSide = (side, buttons, testSection) => {
+  const includesFullWidthButton = areAnyButtonsFullWidth(buttons);
+  const wrapperClassNames = classNames({
+    'flex': includesFullWidthButton,
+    'flex--1': true,
+    [`oui-button-row--${side}`]: side,
+  });
+
   return (
     <div
-      className={ 'flex--1 oui-button-row--' + side }
+      className={ wrapperClassNames }
       data-test-section={ testSection ? testSection + '-' + side : null }>
-      { buttons }
+      { buttons.map((button, index) => {
+        if (includesFullWidthButton) {
+          // `ButtonRow` groups that contain a full width `Button` must wrap
+          // the buttons in `div`'s of equal spacing.
+          const containerClasses = classNames({
+            'push-half--right': index !== (buttons.length - 1),
+            'push-half--left': index !== 0,
+            'flex--1': true,
+          });
+
+          return (
+            <div
+              className={ containerClasses }
+              key={ index }>
+              { button }
+            </div>
+          );
+        }
+        return button;
+      }) }
     </div>
   );
 };
@@ -21,12 +60,15 @@ const ButtonRow = (props) => {
       className='oui-button-row flex'
       data-test-section={ props.testSection }>
       { props.leftGroup ? ButtonRowSide('left', props.leftGroup, props.testSection) : null }
+      { props.centerGroup ? ButtonRowSide('center', props.centerGroup, props.testSection) : null }
       { props.rightGroup ? ButtonRowSide('right', props.rightGroup, props.testSection) : null }
     </div>
   );
 };
 
 ButtonRow.propTypes = {
+  /** Array of buttons for the center */
+  centerGroup: React.PropTypes.arrayOf(React.PropTypes.element),
   /** Array of buttons for left side */
   leftGroup: React.PropTypes.arrayOf(React.PropTypes.element),
   /** Array of buttons for right side */