+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Feature] Add a `centerGroup` prop to `ButtonRow`. (#651)
+
 ### Fixed
 - [Patch] Fix a CSS selector that doesn't work in Enzyme. (#649)